change timing to use nanos instead

+++ src/main/java/com/skillnoob/dh/benchmark/data/BenchmarkResult.java
@@ -1,4 +1,4 @@
 package com.skillnoob.dh.benchmark.data;
 
-public record BenchmarkResult(long elapsedTime, long dbSize, int averageCps) {
+public record BenchmarkResult(long elapsedTime, long dbSize, long averageCps) {
 }
